@@ -14,15 +14,15 @@
     "minimum-stability": "stable",
     "require": {
         "php": ">=5.4.0",
-        "symfony/config": "~2.3|~3.0",
-        "symfony/dependency-injection": "~2.3|~3.0",
-        "symfony/filesystem": "~2.3|~3.0",
-        "symfony/yaml": "~2.3|~3.0",
-        "twig/twig": "~1.16",
+        "symfony/config": "~4.0",
+        "symfony/dependency-injection": "~4.0",
+        "symfony/filesystem": "~4.0",
+        "symfony/yaml": "~4.0",
+        "twig/twig": "~2.5",
         "jakoch/phantomjs-installer": "2.1.1-p08"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0",
+        "phpunit/phpunit": "^7",
         "zendframework/zendpdf": "~2.0",
         "smalot/pdfparser": "~0.9"
     },
