@@ -29,7 +29,7 @@ class TemplateRenderer implements TemplateRendererInterface
      * @access public
      * @param \Twig_Environment $twig
      */
-    public function __construct(\Twig_Environment $twig)
+    public function __construct(\Twig\Environment $twig)
     {
         $this->twig = $twig;
     }
