@@ -15,7 +15,7 @@ use JonnyW\PhantomJs\Procedure\Input;
  *
  * @author Jon Wenmoth <contact@jonnyw.me>
  */
-class InputTest extends \PHPUnit_Framework_TestCase
+class InputTest extends \PHPUnit\Framework\TestCase
 {
 
 /** +++++++++++++++++++++++++++++++++++ **/
