@@ -16,7 +16,7 @@ use JonnyW\PhantomJs\Procedure\ProcedureLoaderFactory;
  *
  * @author Jon Wenmoth <contact@jonnyw.me>
  */
-class ProcedureLoaderFactoryTest extends \PHPUnit_Framework_TestCase
+class ProcedureLoaderFactoryTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * Test directory
