@@ -104,7 +104,9 @@ class Procedure implements ProcedureInterface
                 array('pipe', 'w')
             );
 
-            $process = proc_open(escapeshellcmd(sprintf('%s %s', $this->engine->getCommand(), $executable)), $descriptorspec, $pipes, null, null);
+            $cmd = sprintf('%s %s', $this->engine->getCommand(), $executable);
+            $cmd = escapeshellcmd(str_replace('/', '\\', $cmd));
+            $process = proc_open($cmd, $descriptorspec, $pipes, null, null);
 
             if (!is_resource($process)) {
                 throw new ProcedureFailedException('proc_open() did not return a resource');
@@ -122,7 +124,6 @@ class Procedure implements ProcedureInterface
             $output->import(
                 $this->parser->parse($result)
             );
-
             $this->engine->log($log);
 
             $this->remove($executable);
